// src/types.ts
export type Action = {
    type: string;
    payload: any;
    meta?: {
        skipHistory?: boolean;
        [key: string]: any;
    }
};
export type ToDispatch = Action[];

export interface Operation {
    add: (...actions: Action[]) => void;
    clear: () => void;
    addDiff: (fn: DiffFactoryFn) => void;
    performDiff: (fn: DiffFactoryFn) => void;
    submit: () => void;
    state: ActionState;
}